requirements.txt
Deprecated==1.2.14
img2pdf==0.5.1
lxml==5.3.0
numpy==1.21.6
opencv-python-headless
opencv-contrib-python-headless
packaging==24.0
pdf2img==0.1.2
pikepdf==6.2.9
pillow==9.5.0
pytesseract==0.3.10
regex==2024.4.16
wrapt==1.16.0